requirements.txt
# Nota: ffmpeg debe estar instalado en el sistema operativo
# Descarga de https://ffmpeg.org/download.html y añade al PATH
openai-whisper>=20231117
ffmpeg-python>=0.2.0
googletrans==3.1.0a0
pytube>=15.0.0
pydub>=0.25.1
streamlit>=1.28.0
python-dotenv>=1.0.0